fix: rename file from utils to mockData
